Fully split scans
