+++ general/scans/doc/source/tutorial.rst
@@ -136,7 +136,7 @@ Plot Motor Scan
   :meth:`scans.Spec.rscan` method which defaults to a relative scan,
   instead of an absolute.
 
-  >>> from general.scans.scans.spec import rscan
+  >>> from instrument.larmor.scans import rscan
   >>> rscan(theta, -1, 1, 0.5, 5)
   Taking a count at theta=3.00 and two theta=0.00
   Taking a count at theta=3.50 and two theta=0.00
@@ -446,7 +446,7 @@ SPEC compatibility
   represent a number of frames instead of a monitor count, since
   waiting for a monitor count is currently unsupported.
 
-  >>> from general.scans.scans.spec import ascan, dscan
+  >>> from instrument.larmor.scans import ascan, dscan
   >>> ascan(theta, 0, 2, 10, 1)
   Taking a count at theta=0.00 and two theta=3.00
   Taking a count at theta=0.20 and two theta=3.00
